+++ assets/js/event/rdv/RdvForm.js
@@ -22,11 +22,9 @@ export default class RdvForm {
         this.themeColor = document.getElementById('header').dataset.color
 
         this.modalRdvElt = document.getElementById('modal-rdv')
-        this.modalElt = new Modal(this.modalRdvElt)
+        this.rdvModal = new Modal(this.modalRdvElt)
 
         new SearchLocation('rdv_search_location')
-
-        this.formValidator = new FormValidator(this.modalRdvElt)
 
         this.btnAddAlertElt = document.querySelector('button[data-add-widget]')
         this.btnSaveRdvElt = this.modalRdvElt.querySelector('button[data-action="save-rdv"]')
@@ -60,16 +58,16 @@ export default class RdvForm {
 
         this.currentUserId = document.getElementById('user-name').dataset.userId
 
-        const eventObject = {name: 'onModal', elementId: 'modal-rdv'}
-        this.usersSelectManager = new SelectManager('#rdv_users', eventObject, {width: '100%'})
-        this.tagsSelectManager = new SelectManager('#rdv_tags', eventObject)
+        this.formValidator = new FormValidator(this.formRdvElt)
+        this.usersSelectManager = new SelectManager('#rdv_users')
+        this.tagsSelectManager = new SelectManager('#rdv_tags')
 
         this.alertsCollectionManager = new WidgetCollectionManager(this.afterToAddAlert.bind(this), null, 3)
 
         this.editColumnRdvElt = document.querySelector('table#table-rdvs th[data-path-edit-rdv]')
         this.editContainRdvElt = document.querySelector('div.calendar-table div[data-path-edit-rdv]')
 
-        this.updateModalElt = new Modal(document.getElementById('modal-update'))
+        this.updateModal = new Modal(document.getElementById('modal-update'))
 
         this.googleCalendarCheckbox = this.modalRdvElt.querySelector('input[name="rdv[_googleCalendar]"]')
 
@@ -132,11 +130,11 @@ export default class RdvForm {
         this.btnDeleteRdvElt.classList.add('d-none')
 
         if (e !== undefined && (e.target.className && e.target.className.search('calendar-event') !== 0)) {
-            this.modalElt.show()
-            this.tagsSelectManager.clearSelect()
+            this.rdvModal.show()
+            this.tagsSelectManager.clearItems()
         }
 
-        this.usersSelectManager.updateSelect(this.currentUserId)
+        this.usersSelectManager.updateItems(this.currentUserId)
 
         this.resetAlerts()
     }
@@ -251,11 +249,11 @@ export default class RdvForm {
 
         const tagsIds = []
         rdv.tags.forEach(tags => tagsIds.push(tags.id))
-        this.tagsSelectManager.updateSelect(tagsIds)
+        this.tagsSelectManager.updateItems(tagsIds)
 
         const userIds = []
         rdv.users.forEach(user => userIds.push(user.id))
-        this.usersSelectManager.updateSelect(userIds)
+        this.usersSelectManager.updateItems(userIds)
 
         this.supportSelectElt.value = ''
         this.supportSelectElt.disabled = rdv.supportGroup !== null
@@ -292,7 +290,7 @@ export default class RdvForm {
 
         this.initAlerts(rdv)
 
-        this.modalElt.show()
+        this.rdvModal.show()
     }
 
     /**
@@ -344,7 +342,7 @@ export default class RdvForm {
             || (rdvModel.isDifferent(this.rdvBeforeUpdate) && (this.googleCalendarCheckbox.checked
                 || this.outlookCalendarCheckbox.checked))
         ) {
-            this.updateModalElt.show()
+            this.updateModal.show()
 
             const listApis = () => {
                 let list = {}
@@ -389,6 +387,8 @@ export default class RdvForm {
             this.usersSelecElt,
         ]
 
+        this.formRdvElt.classList.add('was-validated')
+
         document.querySelector('#alerts-fields-list').querySelectorAll('input, select').forEach(fieldElt => fieldElts.push(fieldElt))
 
         fieldElts.forEach(fieldElt => {
@@ -426,6 +426,6 @@ export default class RdvForm {
     closeModal() {
         // The ordre is important
         document.getElementById('js-btn-cancel').click()
-        this.modalElt.hide()
+        this.rdvModal.hide()
     }
 }