+++ src/Form/Evaluation/EvalAdmPersonType.php
@@ -102,7 +102,10 @@ class EvalAdmPersonType extends AbstractType
             ->add('cndaId', null, [
                 'attr' => ['data-mask-type' => 'number'],
             ])
-            ->add('_country', CountryType::class)
+            ->add('_country', CountryType::class, [
+                'placeholder' => 'placeholder.select',
+                'mapped' => false,
+            ])
         ;
 
         if (!$builder->get('nationality')->getData()) {