+++ src/Form/Event/RdvSearchType.php
@@ -15,6 +15,7 @@ use Doctrine\Common\Collections\ArrayCollection;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
+use Symfony\Component\Form\Extension\Core\Type\SearchType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Form\FormEvent;
 use Symfony\Component\Form\FormEvents;
@@ -40,21 +41,21 @@ class RdvSearchType extends AbstractType
         $this->setFormData($builder);
 
         $builder
-            ->add('id', null, [
+            ->add('id', SearchType::class, [
                 'label_attr' => ['class' => 'sr-only'],
                 'attr' => [
                     'placeholder' => 'ID',
                     'class' => 'w-max-80',
                 ],
             ])
-            ->add('title', null, [
+            ->add('title', SearchType::class, [
                 'label_attr' => ['class' => 'sr-only'],
                 'attr' => [
                     'placeholder' => 'Title',
                     'class' => 'w-max-170',
                 ],
             ])
-            ->add('fullname', null, [
+            ->add('fullname', SearchType::class, [
                 'label_attr' => ['class' => 'sr-only'],
                 'attr' => [
                     'placeholder' => 'search.fullname.placeholder',
@@ -65,7 +66,7 @@ class RdvSearchType extends AbstractType
                 'multiple' => true,
                 'choices' => Choices::getChoices(Rdv::STATUS),
                 'attr' => [
-                    'class' => 'multi-select',
+                    'class' => 'multi-select w-max-220',
                     'placeholder' => 'placeholder.status',
                     'size' => 1,
                 ],
@@ -85,7 +86,7 @@ class RdvSearchType extends AbstractType
                     return $repo->findUsersOfCurrentUserQueryBuilder($this->user);
                 },
                 'attr' => [
-                    'class' => 'multi-select w-min-150 w-max-220',
+                    'class' => 'multi-select w-min-180 w-max-220',
                     'placeholder' => 'event.users.placeholder',
                     'size' => 1,
                 ],
@@ -100,7 +101,7 @@ class RdvSearchType extends AbstractType
                 'choice_label' => 'name',
                 'label_attr' => ['class' => 'sr-only'],
                 'attr' => [
-                    'class' => 'multi-select w-min-200 w-max-220',
+                    'class' => 'multi-select w-max-220',
                     'placeholder' => 'placeholder.tags',
                     'size' => 1,
                 ],