+++ src/Form/Event/TaskSearchType.php
@@ -15,6 +15,7 @@ use Doctrine\Common\Collections\ArrayCollection;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
+use Symfony\Component\Form\Extension\Core\Type\SearchType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Form\FormEvent;
 use Symfony\Component\Form\FormEvents;
@@ -39,21 +40,21 @@ class TaskSearchType extends AbstractType
         $this->setFormData($builder);
 
         $builder
-            ->add('id', null, [
+            ->add('id', SearchType::class, [
                 'label_attr' => ['class' => 'sr-only'],
                 'attr' => [
                     'placeholder' => 'ID',
                     'class' => 'w-max-80',
                 ],
             ])
-            ->add('title', null, [
+            ->add('title', SearchType::class, [
                 'label_attr' => ['class' => 'sr-only'],
                 'attr' => [
                     'placeholder' => 'search',
                     'class' => 'w-max-170',
                 ],
             ])
-            ->add('fullname', null, [
+            ->add('fullname', SearchType::class, [
                 'label_attr' => ['class' => 'sr-only'],
                 'attr' => [
                     'placeholder' => 'search.fullname.placeholder',
@@ -71,7 +72,7 @@ class TaskSearchType extends AbstractType
                 'multiple' => true,
                 'choices' => Choices::getChoices(Task::LEVEL),
                 'attr' => [
-                    'class' => 'multi-select w-min-140',
+                    'class' => 'multi-select w-max-220',
                     'size' => 1,
                     'placeholder' => 'task.level.placeholder',
                 ],
@@ -81,7 +82,7 @@ class TaskSearchType extends AbstractType
                 'multiple' => true,
                 'choices' => Choices::getChoices(Task::STATUS),
                 'attr' => [
-                    'class' => 'multi-select w-min-160',
+                    'class' => 'multi-select w-max-220',
                     'placeholder' => 'placeholder.status',
                     'size' => 1,
                 ],
@@ -104,7 +105,7 @@ class TaskSearchType extends AbstractType
                     return $repo->findUsersOfCurrentUserQueryBuilder($this->user);
                 },
                 'attr' => [
-                    'class' => 'multi-select w-min-150 w-max-220',
+                    'class' => 'multi-select w-max-220',
                     'placeholder' => 'event.users.placeholder',
                     'size' => 1,
                 ],
@@ -119,7 +120,7 @@ class TaskSearchType extends AbstractType
                 'choice_label' => 'name',
                 'label_attr' => ['class' => 'sr-only'],
                 'attr' => [
-                    'class' => 'multi-select w-min-200 w-max-220',
+                    'class' => 'multi-select w-max-220',
                     'placeholder' => 'placeholder.tags',
                     'size' => 1,
                 ],