+++ src/Form/Organization/Place/PlaceGroupHotelType.php
@@ -32,7 +32,7 @@ class PlaceGroupHotelType extends AbstractType
                         return $repo->getPlacesQueryBuilder($service);
                     },
                     'label' => Service::SERVICE_TYPE_HOTEL === $service->getType() ? 'hotelName' : 'place.name',
-                    'attr' => ['data-select' => 'advanced'],
+                    'attr' => ['autocomplete' => true],
                     'placeholder' => 'placeholder.select',
                     'required' => true,
                 ])