+++ src/Form/Organization/Service/ServiceSearchType.php
@@ -8,6 +8,7 @@ use App\Form\Utils\Choices;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
+use Symfony\Component\Form\Extension\Core\Type\SearchType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
@@ -16,7 +17,7 @@ class ServiceSearchType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
-            ->add('name', null, [
+            ->add('name', SearchType::class, [
                 'label_attr' => [
                     'class' => 'sr-only',
                 ],
@@ -26,7 +27,7 @@ class ServiceSearchType extends AbstractType
                     'autocomplete' => 'off',
                 ],
             ])
-            ->add('city', null, [
+            ->add('city', SearchType::class, [
                 'label_attr' => [
                     'class' => 'sr-only',
                 ],