+++ src/Form/Support/Document/DocumentSearchType.php
@@ -9,6 +9,7 @@ use App\Form\Type\ServiceDeviceReferentSearchType;
 use App\Repository\Organization\TagRepository;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\SearchType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
@@ -24,13 +25,13 @@ class DocumentSearchType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
-            ->add('id', null, [
+            ->add('id', SearchType::class, [
                 'attr' => [
                     'placeholder' => 'ID',
                     'class' => 'w-max-80',
                 ],
             ])
-            ->add('name', null, [
+            ->add('name', SearchType::class, [
                 'attr' => ['placeholder' => 'Search'],
             ])
             ->add('date', DateSearchType::class, [
@@ -48,7 +49,7 @@ class DocumentSearchType extends AbstractType
                 'choice_label' => 'name',
                 'label_attr' => ['class' => 'sr-only'],
                 'attr' => [
-                    'class' => 'multi-select w-min-200 w-max-220',
+                    'class' => 'multi-select w-max-220',
                     'placeholder' => 'placeholder.tags',
                     'size' => 1,
                 ],